Refactored code to use one activity and employ dynamic changing of layout view.
