+++ app/src/main/AndroidManifest.xml
@@ -7,7 +7,7 @@
         android:required="true" />
 
     <uses-permission android:name="android.permission.INTERNET" />
-    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
 
     <application
         android:allowBackup="true"
@@ -23,8 +23,15 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity android:name=".SecondActivity"/>
+        <activity android:name=".SecondActivity" />
+        <activity
+            android:name=".Demotivational"
+            android:label="@string/title_activity_demotivational" >
+        </activity>
+        <activity
+            android:name=".Vanilla"
+            android:label="@string/title_activity_vanilla" >
+        </activity>
     </application>
 
 </manifest>
